Create Code with proper comments here.• To demonstrate your understanding of problem-solving using C/C++ concepts
• To demonstrate your understanding of OS concepts:

• Processes and their Execution and Context switching
• Threads
• Synchronization Schemes
• Process/Threads Scheduling
• Memory Management

• To demonstrate your understanding of reading and writing using IPC methods

• Pipes (Ordinary and named pipes)
• Shared Memory
Project Details

Project Name:
Potion Explosion – Game
Projection Description:

Potion Explosion OS is a multiplayer game designed to simulate the essence of potion-crafting and marble-
crashing from the original Potion Explosion game. The game will be developed as an application running on

an operating system, leveraging various OS concepts to manage game resources, processes, and interactions
between players.

Page 5 of 7
Game Details

Game Components:
• Dispenser & Balls (Red, Yellow, Blue, Black)

• Potion Tiles

• Desk board

Page 6 of 7

Game Rules:
Note that you should see the rules of the game from rulebook that I provided as anattachment and I
am also providing How to Play Video link
(https://www.youtube.com/watch?v=iSODqRFNe3I ) for your convenience
Also, the following things are not included in game that are mentioned in the rulebook.
• Potion Effects
• Skill Token
• Little Help Tokens
• Not need to make Potion Tiles with Star.
• No need to make Ingredient Pool in Potion TilesMake
only 5 styles of Potion Tiles.

There are bonus marks of adding more feature (that are not included) and make game usinggraphics.
You have to make this game for two players. At the end of the game also show the winner and scores
of the player. If you are making this game in console then you have to think how can you display the
dispenser with balls and potion tiles of the players.